+++ zap/src/main/java/org/zaproxy/zap/extension/ascan/filters/GenericFilterData.java
@@ -17,7 +17,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.zaproxy.zap.scan.filters;
+package org.zaproxy.zap.extension.ascan.filters;
 
 import java.util.Collection;
 import java.util.LinkedHashSet;
