+++ zap/src/main/java/org/zaproxy/zap/extension/ascan/filters/impl/HttpStatusCodeScanFilter.java
@@ -17,27 +17,26 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.zaproxy.zap.scan.filters.impl;
+package org.zaproxy.zap.extension.ascan.filters.impl;
 
-import java.util.List;
 import org.parosproxy.paros.Constant;
 import org.parosproxy.paros.model.HistoryReference;
+import org.zaproxy.zap.extension.ascan.filters.FilterResult;
 import org.zaproxy.zap.model.StructuralNode;
-import org.zaproxy.zap.scan.filters.FilterResult;
 
 /** @author KSASAN preetkaran20@gmail.com */
-public class TagScanFilter extends AbstractScanFilter<String> {
+public class HttpStatusCodeScanFilter extends AbstractScanFilter<Integer> {
 
-    private static final String FILTER_TYPE = "scan.filter.filterType.Tag";
+    private static final String FILTER_TYPE = "scan.filter.filterType.HttpStatusCode";
 
     @Override
     public FilterResult isFiltered(StructuralNode node) {
         HistoryReference href = node.getHistoryReference();
         if (href != null) {
-            List<String> nodeTags = href.getTags();
-            return this.isFiltered(nodeTags);
+            int statusCode = href.getStatusCode();
+            return this.isFiltered(statusCode);
         } else {
-            return FilterResult.FILTERED_RESULT;
+            return FilterResult.NOT_FILTERED;
         }
     }
