+++ zap/src/main/resources/org/zaproxy/zap/resources/Messages.properties
@@ -451,13 +451,13 @@ ascan.options.strength.insane.label = (MUCH more attacks)
 ascan.options.strengthTo.label		= Strength To
 ascan.options.thresholdTo.label		= Threshold To
 
-scan.filter.filtercriteria.include = Request should include one of the following {0} values :- {1} 
-scan.filter.filtercriteria.exclude = Request should not include the following {0} value :- {1}  
-scan.filter.filtercriteria.includeall = Request should include all of the following {0} values :- {1}
-scan.filter.filterType.Tag = Tag; 
-scan.filter.filterType.HttpMethod = HTTP Method;
-scan.filter.filterType.HttpStatusCode = HTTP Status Code;
-scan.filter.filterType.URLPattern = URL Pattern;
+scan.filter.filtercriteria.include = Node should include one of the following {0} values: {1} 
+scan.filter.filtercriteria.exclude = Node should not include the following {0} value: {1}  
+scan.filter.filtercriteria.includeall = Node should include all of the following {0} values: {1}
+scan.filter.filterType.Tag = Tag
+scan.filter.filterType.HttpMethod = HTTP Method
+scan.filter.filterType.HttpStatusCode = HTTP Status Code
+scan.filter.filterType.URLPattern = URL Pattern
 
 # Prob need to move ;)
 ascan.policymgr.button.add		= Add
